pyproject.toml
[project]
name = "pdf-extractor"
version = "0.1.0"
description = "Convert PDF files to Markdown format"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymupdf4llm>=0.0.17",
    # RAG Components
    "sentence-transformers>=2.2.2",
    "chromadb>=0.4.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    # LLM Integration
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
    "langchain-ollama>=0.1.0",
    # MCP Server
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    # Utilities
    "numpy>=1.24.0",
    "tiktoken>=0.5.0",
    "pyyaml>=6.0",
    # HTTP Server
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=5.0.0",
    "mypy>=1.13.0",
    "ruff>=0.8.5",
]

[project.scripts]
pdf-extractor = "src.pdf_extractor:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_extractor", "src/rag_engine", "src/mcp"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]